+++ core/src/test/java/uk/gov/justice/services/core/it/LegacySenderIT.java
@@ -22,7 +22,6 @@ import uk.gov.justice.services.core.dispatcher.RequesterProducer;
 import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
 import uk.gov.justice.services.core.dispatcher.SystemUserProvider;
 import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
-import uk.gov.justice.services.core.eventbuffer.PassThroughEventBufferService;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
 import uk.gov.justice.services.core.it.util.sender.RecordingSender;
@@ -86,7 +85,6 @@ public class LegacySenderIT {
             RequesterProducer.class,
             SenderProducer.class,
             ServiceComponentObserver.class,
-            PassThroughEventBufferService.class,
             LoggerProducer.class,
             TestSystemUserProvider.class,
             SystemUserUtil.class,