+++ generators/messaging-adapter-generator/src/test/java/uk/gov/justice/raml/jms/it/JmsAdapterToHandlerIT.java
@@ -29,9 +29,7 @@ import uk.gov.justice.services.core.dispatcher.RequesterProducer;
 import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
 import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
 import uk.gov.justice.services.core.enveloper.Enveloper;
-import uk.gov.justice.services.core.eventbuffer.EventBufferInterceptor;
 import uk.gov.justice.services.core.eventfilter.AllowAllEventFilter;
-import uk.gov.justice.services.core.eventfilter.EventFilterInterceptor;
 import uk.gov.justice.services.core.extension.AnnotationScanner;
 import uk.gov.justice.services.core.extension.BeanInstantiater;
 import uk.gov.justice.services.core.interceptor.InterceptorCache;
@@ -44,7 +42,9 @@ import uk.gov.justice.services.core.json.JsonSchemaLoader;
 import uk.gov.justice.services.core.json.JsonSchemaValidator;
 import uk.gov.justice.services.core.sender.ComponentDestination;
 import uk.gov.justice.services.core.sender.SenderProducer;
+import uk.gov.justice.services.event.buffer.EventBufferInterceptor;
 import uk.gov.justice.services.event.buffer.api.EventBufferService;
+import uk.gov.justice.services.event.filter.EventFilterInterceptor;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.messaging.JsonObjectEnvelopeConverter;
 import uk.gov.justice.services.messaging.jms.DefaultJmsEnvelopeSender;
@@ -129,7 +129,6 @@ public class JmsAdapterToHandlerIT extends AbstractJmsAdapterGenerationIT {
             EmptySystemUserProvider.class,
             SystemUserUtil.class,
             BeanInstantiater.class
-
 
 
     })